fix: changed datatype in counter

+++ src/sort/insertion_sort.rs
@@ -2,7 +2,7 @@
 #[allow(unused_variables)]
 pub fn insertion_sort(arr: &mut [u8]) {
 	for i in 1..arr.len() {
-		let mut j: u32 = i as u32;
+		let mut j: u16 = i as u16;
 
 		while j > 0 && arr[j as usize] < arr[(j - 1) as usize] {
 			let temp = arr[j as usize];
